Rename : package structure modify
