+++ src/main/java/bitcopark/library/repository/Book/BookRequestRepository.java
@@ -1,7 +1,6 @@
-package bitcopark.library.repository.Member;
+package bitcopark.library.repository.Book;
 
 import bitcopark.library.entity.member.BookRequest;
-import bitcopark.library.repository.Book.BookRepositoryCustom;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BookRequestRepository extends JpaRepository<BookRequest, Long>, BookRequestRepositoryCustom {
