+++ src/main/java/bitcopark/library/repository/Book/BookRequestRepositoryCustom.java
@@ -1,10 +1,8 @@
-package bitcopark.library.repository.Member;
+package bitcopark.library.repository.Book;
 
-import bitcopark.library.service.Member.BookRequestPageDto;
+import bitcopark.library.service.Book.BookRequestPageDto;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-
-import java.util.List;
 
 public interface BookRequestRepositoryCustom {
