+++ src/main/java/bitcopark/library/service/Book/BookRequestDetailDto.java
@@ -1,0 +1,4 @@
+package bitcopark.library.service.Book;
+
+public class BookRequestDetailDto {
+}
