+++ src/main/java/bitcopark/library/temp/tempInitGenerate.java
@@ -9,14 +9,13 @@ import bitcopark.library.entity.member.BookRequestApprove;
 import bitcopark.library.entity.member.Member;
 import bitcopark.library.entity.member.MemberGender;
 import bitcopark.library.repository.Board.CategoryRepository;
-import bitcopark.library.repository.Book.BookRepository;
 import bitcopark.library.repository.Member.MemberRepository;
 import bitcopark.library.service.Board.CategoryService;
 import bitcopark.library.service.Book.BookBorrowService;
 import bitcopark.library.service.Book.BookLikeService;
 import bitcopark.library.service.Book.BookReservationService;
 import bitcopark.library.service.Book.BookService;
-import bitcopark.library.service.Member.BookRequestService;
+import bitcopark.library.service.Book.BookRequestService;
 import bitcopark.library.service.Member.MemberService;
 import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
